+++ patches/src/main/kotlin/dev/jkcarino/revanced/patches/reddit/ad/Fingerprints.kt
@@ -1,4 +1,4 @@
-package dev.jkcarino.revanced.patches.reddit
+package dev.jkcarino.revanced.patches.reddit.ad
 
 import app.revanced.patcher.fingerprint
 import com.android.tools.smali.dexlib2.AccessFlags
@@ -36,22 +36,19 @@ internal val interceptFingerprint = fingerprint {
 
 internal val okHttpConstructorFingerprint = fingerprint {
     accessFlags(AccessFlags.PUBLIC, AccessFlags.CONSTRUCTOR)
-    parameters("Lokhttp3/OkHttpClient\$Builder;")
+    parameters("L")
     opcodes(
         Opcode.MOVE_OBJECT_FROM16,
         Opcode.MOVE_OBJECT_FROM16,
         Opcode.CONST_STRING,
         Opcode.INVOKE_STATIC,
-        Opcode.INVOKE_DIRECT_RANGE,
-        Opcode.INVOKE_VIRTUAL_RANGE,
+        null,
+        null,
         Opcode.MOVE_RESULT_OBJECT,
         Opcode.IPUT_OBJECT,
-        Opcode.INVOKE_VIRTUAL_RANGE,
+        null,
         Opcode.MOVE_RESULT_OBJECT,
     )
-    custom { method, classDef ->
-        classDef.type == "Lokhttp3/OkHttpClient;"
-    }
 }
 
 internal val realBufferedSourceCommonIndexOfFingerprint = fingerprint {